@@ -1,14 +1,14 @@
 {
   "name": "reindeer",
   "description": "Provides schema enforcement based on Elasticsearch type mappings",
-  "main": "./lib/index.js",
-  "version": "2.1.3",
+  "main": "./dist",
+  "version": "3.0.0",
   "author": "Joshua Thomas (https://github.com/brozeph)",
   "contributors": [
     "Robin Momii (https://github.com/rmomii)"
   ],
   "engines": {
-    "node": ">= 4.0"
+    "node": ">= 6.0"
   },
   "keywords": [
     "elasticsearch",
@@ -18,24 +18,58 @@
   ],
   "repository": "git://github.com/brozeph/reindeer.git",
   "homepage": "http://github.com/brozeph/reindeer",
-  "dependencies": {
-    "es": "^0.6.0",
-    "lodash.clonedeep": "^4.5.0"
-  },
-  "devDependencies": {
-    "chai": "^3.2.0",
-    "del": "^1.2.1",
-    "gulp": "^3.9.0",
-    "gulp-coveralls": "^0.1.4",
-    "gulp-eslint": "^2.0.0",
-    "gulp-istanbul": "^0.10.0",
-    "gulp-mocha": "^2.1.3",
-    "gulp-util": "^3.0.7",
-    "nock": "^2.11.0",
-    "uuid": "^2.0.1"
+  "nyc": {
+    "all": true,
+    "exclude": [
+      "src/index.js"
+    ],
+    "include": [
+      "src"
+    ],
+    "instrument": false,
+    "report-dir": "./reports",
+    "reporter": [
+      "lcov",
+      "text",
+      "text-summary"
+    ],
+    "require": [
+      "@babel/register"
+    ],
+    "sourceMap": false
   },
   "license": "MIT",
   "scripts": {
-    "test": "gulp lint && gulp test-integration && gulp coveralls"
+    "lint": "gulp lint",
+    "posttest": "nyc report --reporter=text-lcov | coveralls",
+    "prepublish": "gulp build",
+    "pretest": "gulp clean && gulp lint",
+    "test": "NODE_ENV=test nyc mocha ./test/src",
+    "test:unit": "NODE_ENV=test nyc mocha ./test/src"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.3.3",
+    "@babel/plugin-transform-runtime": "^7.2.0",
+    "@babel/preset-env": "^7.3.1",
+    "@babel/register": "^7.0.0",
+    "babel-eslint": "^10.0.1",
+    "babel-plugin-istanbul": "^5.1.1",
+    "chai": "^4.2.0",
+    "coveralls": "^3.0.2",
+    "eslint-plugin-babel": "^5.3.0",
+    "gulp": "^4.0.0",
+    "gulp-babel": "^8.0.0",
+    "gulp-clean": "^0.4.0",
+    "gulp-eslint": "^5.0.0",
+    "gulp-sourcemaps": "^2.6.5",
+    "mocha": "^6.0.0",
+    "nock": "^10.0.6",
+    "nyc": "^13.3.0",
+    "uuid": "^3.3.2"
+  },
+  "dependencies": {
+    "@babel/runtime-corejs2": "^7.3.1",
+    "es": "^0.6.0"
   }
 }
